Create Firmware Folder
Arduino IDE will not load OpeNITHM.ino unless it's in a folder named OpeNITHM
