+++ Firmware/OpeNITHM/CapacitiveSensor.cpp
